# docs/swagger.yaml
openapi: 3.0.4
info:
  title: PPO Project.HttpServer
  version: v1
paths:
...
  '/api/v1/positions/{positionId}':
    get:
      tags:
        - Position
      parameters:
        - name: positionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PositionDto'
            application/json:
              schema:
                $ref: '#/components/schemas/PositionDto'
            text/json:
              schema:
                $ref: '#/components/schemas/PositionDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    delete:
      tags:
        - Position
      parameters:
        - name: positionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
          content:
            text/plain:
              schema:
                type: boolean
            application/json:
              schema:
                type: boolean
            text/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  /api/v1/positions:
    post:
      tags:
        - Position
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePositionDto'
          text/json:
            schema:
              $ref: '#/components/schemas/CreatePositionDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreatePositionDto'
        required: true
      responses:
        '201':
          description: Created
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PositionDto'
            application/json:
              schema:
                $ref: '#/components/schemas/PositionDto'
            text/json:
              schema:
                $ref: '#/components/schemas/PositionDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
  '/api/v1/employees/{employeeId}/positionHistories/{positionId}':
    get:
      tags:
        - PositionHistory
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: positionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PositionHistoryDto'
            application/json:
              schema:
                $ref: '#/components/schemas/PositionHistoryDto'
            text/json:
              schema:
                $ref: '#/components/schemas/PositionHistoryDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    delete:
      tags:
        - PositionHistory
      parameters:
        - name: positionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: No Content
          content:
            text/plain:
              schema:
                type: boolean
            application/json:
              schema:
                type: boolean
            text/json:
              schema:
                type: boolean
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
    patch:
      tags:
        - PositionHistory
      parameters:
        - name: positionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePositionHistoryDto'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdatePositionHistoryDto'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdatePositionHistoryDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/PositionHistoryDto'
            application/json:
              schema:
                $ref: '#/components/schemas/PositionHistoryDto'
            text/json:
              schema:
                $ref: '#/components/schemas/PositionHistoryDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
  '/api/v1/employees/{employeeId}/positionHistories':
    get:
      tags:
        - PositionHistory
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            default: 1
            nullable: false
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            enum: [10, 20, 40]
            default: 10
            nullable: false
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PositionHistoryDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PositionHistoryDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PositionHistoryDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  '/api/v1/employees/{employeeId}/subordinates/positionHistories':
    get:
      tags:
        - PositionHistory
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
            nullable: true
        - name: pageNumber
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            default: 1
            nullable: false
        - name: pageSize
          in: query
          required: true
          schema:
            type: integer
            enum: [10, 20, 40]
            default: 10
            nullable: false
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PositionHistoryDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PositionHistoryDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PositionHistoryDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
  '/api/v1/employees/{headEmployeeId}/currentSubordinates/positionHistories':
    get:
      tags:
        - PositionHistory
      parameters:
        - name: headEmployeeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PositionHierarchyWithEmployeeDto'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PositionHierarchyWithEmployeeDto'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PositionHierarchyWithEmployeeDto'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
        '500':
          description: Internal Server Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
            text/json:
              schema:
                $ref: '#/components/schemas/ErrorDto'
...
components:
  schemas:
...
    CreatePositionDto:
      required:
        - title
        - companyId
        - parentId
      type: object
      properties:
        parentId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          nullable: false
        companyId:
          type: string
          format: uuid
          nullable: false
      additionalProperties: false
...
    ErrorDto:
      required:
        - errorType
        - message
      type: object
      properties:
        errorType:
          type: string
          nullable: false
        message:
          type: string
          nullable: true
      additionalProperties: false
...
    PositionDto:
      required:
        - id
        - title
        - companyId
        - parentId
        - isDeleted
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: false
        parentId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          nullable: false
        companyId:
          type: string
          format: uuid
          nullable: false
        isDeleted:
          type: boolean
          nullable: false
      additionalProperties: false
...
    PositionHierarchyWithEmployeeDto:
      required:
        - positionId
        - employeeId
        - title
        - level
        - parentId
      type: object
      properties:
        positionId:
          type: string
          format: uuid
          nullable: false
        employeeId:
          type: string
          format: uuid
          nullable: false
        parentId:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
          nullable: false
        level:
          type: integer
          format: int32
          nullable: false
      additionalProperties: false
    PositionHistoryDto:
      required:
        - positionId
        - employeeId
        - startDate
        - endDate
      type: object
      properties:
        positionId:
          type: string
          format: uuid
          nullable: false
        employeeId:
          type: string
          format: uuid
          nullable: false
        startDate:
          type: string
          format: date
          nullable: false
        endDate:
          type: string
          format: date
          nullable: true
      additionalProperties: false
...
    UpdatePositionHistoryDto:
      required:
        - startDate
        - endDate
      type: object
      properties:
        startDate:
          type: string
          format: date
          nullable: true
        endDate:
          type: string
          format: date
          nullable: true
      additionalProperties: false
...
  securitySchemes:
    Bearer:
      type: apiKey
      name: Authorization
      in: header
security:
  - Bearer: [ ]